Cleaned up the main script considerably.
The anfo> prompt and pageviewer> prompt are now in separate files,
and all commands they call are in the file 'commands.py'.
The main script to run is now 'anfo.py'.

The commands themselves are cleaned up a little, too.  For one,
the docstrings are improved to be more explicit about arguments.
Printing error messages has now been delegated to the UI loops, which
just print messages attached to exceptions.

The pageviewer is slightly better in how it handles query
calls.  Rather than calling the pageviewer again, instead just the list
of songs and certain variables are updated.  This _might_ get rid of the
issue where there was significant lag in quitting after being idle for
a while, which is perhaps caused by lists of songs being loaded from
hard disk when going back through previous invocations of the pageviewer.